+++ lib/repos.txt
@@ -1,0 +1,3 @@
+GoogleCloudPlatform/google-cloud-node
+GoogleCloudPlatform/google-cloud-java
+GoogleCloudPlatform/google-cloud-dotnet
