+++ package.json
@@ -20,15 +20,15 @@
   },
   "preferGlobal": true,
   "bin": {
-    "labelcat": "cli.js"
+    "labelcat": "src/cli.js"
   },
   "engines": {
     "node": ">= 8.x"
   },
   "scripts": {
     "cover": "nyc --reporter=lcov mocha test/*.js && nyc report",
-    "lint": "eslint src/ samples/ system-test/ test/",
-    "prettier": "prettier --write src/*.js src/*/*.js test/*.js test/*/*.js system-test/*.js system-test/*/*.js",
+    "lint": "eslint src/ system-test/ test/",
+    "prettier": "prettier --write src/*.js test/*.js system-test/*.js",
     "system-test": "mocha system-test/*.js --timeout 600000",
     "test-no-cover": "mocha test/*.js",
     "test": "npm run cover"
@@ -38,6 +38,7 @@
     "axios": "^0.18.0",
     "csv-write-stream": "^2.0.0",
     "json2csv": "^4.2.1",
+    "proxyquire": "^2.1.0",
     "yargs": "^12.0.2"
   },
   "devDependencies": {
